+++ tests/test_pillar_model.py
@@ -78,63 +78,6 @@ def test_pillar_boundary_heights_square():
         Profile.diamond(height=5, width=10, corner_radius=1)
     ]
 )
-def test_pillar_sections_equidistant(p: Union[PillarProfile, Profile]):
-    try:
-        pyroll.pillar_model.Config.PILLAR_COUNT = 4
-        pyroll.pillar_model.Config.PILLAR_TYPE = "EQUIDISTANT"
-        fig: plt.Figure = plt.figure()
-        ax: plt.Axes = fig.subplots()
-
-        ax.set_aspect("equal")
-
-        for c in reversed(p.pillar_sections):
-            z, y = np.array(c.boundary.xy)
-            plt.fill(np.concatenate([-z, z]), np.concatenate([y, y]), alpha=0.5)
-
-        plt.plot(*p.cross_section.boundary.xy, c="k")
-        plt.show()
-    finally:
-        del pyroll.pillar_model.Config.PILLAR_COUNT
-        del pyroll.pillar_model.Config.PILLAR_TYPE
-
-
-@pytest.mark.parametrize(
-    "p", [
-        Profile.round(radius=10),
-        Profile.square(side=10, corner_radius=1),
-        Profile.box(height=10, width=5, corner_radius=1),
-        Profile.diamond(height=5, width=10, corner_radius=1)
-    ]
-)
-def test_pillar_sections_uniform(p: Union[PillarProfile, Profile]):
-    try:
-        pyroll.pillar_model.Config.PILLAR_COUNT = 4
-        pyroll.pillar_model.Config.PILLAR_TYPE = "UNIFORM"
-
-        fig: plt.Figure = plt.figure()
-        ax: plt.Axes = fig.subplots()
-
-        ax.set_aspect("equal")
-
-        for c in reversed(p.pillar_sections):
-            z, y = np.array(c.boundary.xy)
-            plt.fill(np.concatenate([-z, z]), np.concatenate([y, y]), alpha=0.5)
-
-        plt.plot(*p.cross_section.boundary.xy, c="k")
-        plt.show()
-    finally:
-        del pyroll.pillar_model.Config.PILLAR_COUNT
-        del pyroll.pillar_model.Config.PILLAR_TYPE
-
-
-@pytest.mark.parametrize(
-    "p", [
-        Profile.round(radius=10),
-        Profile.square(side=10, corner_radius=1),
-        Profile.box(height=10, width=5, corner_radius=1),
-        Profile.diamond(height=5, width=10, corner_radius=1)
-    ]
-)
 def test_uniform_pillar_sections_area(p: Union[PillarProfile, Profile]):
     pyroll.pillar_model.Config.PILLAR_COUNT = 4
     pyroll.pillar_model.Config.PILLAR_TYPE = "UNIFORM"
@@ -156,6 +99,7 @@ def test_equidistant_pillar_widths(p: Union[PillarProfile, Profile]):
 
     assert np.all(np.isclose(p.pillar_widths, p.pillar_widths[0], rtol=1e-3))
 
+
 @pytest.mark.parametrize(
     "p", [
         Profile.round(radius=10),
@@ -169,6 +113,7 @@ def test_equidistant_pillar_boundary_is_profile_boundary_(p: Union[PillarProfile
     pyroll.pillar_model.Config.PILLAR_TYPE = "EQUIDISTANT"
 
     assert np.isclose(p.pillar_boundaries[-1], p.width / 2, rtol=1e-3)
+
 
 @pytest.mark.parametrize(
     "p", [
