+++ tests/test_solve_round_constricted_box.py
@@ -11,13 +11,11 @@ def pillar_spreads(self: RollPass.DiskElement):
     return self.pillar_draughts ** -0.5
 
 
-DISK_ELEMENT_COUNT = 15
-pyroll.pillar_model.Config.PILLAR_COUNT = 30
+def test_solve_round_constricted_box_equidistant(tmp_path: Path, caplog, monkeypatch):
 
-
-def test_solve_round_constricted_box_equidistant(tmp_path: Path, caplog):
     caplog.set_level(logging.INFO, logger="pyroll")
-    pyroll.pillar_model.Config.PILLAR_TYPE = "EQUIDISTANT"
+    monkeypatch.setenv("PILLAR_TYPE", "EQUIDISTANT")
+    monkeypatch.setenv("PILLAR_COUNT", 30)
 
     in_profile = Profile.round(
         diameter=19.5e-3,
@@ -48,7 +46,7 @@ def test_solve_round_constricted_box_equidistant(tmp_path: Path, caplog):
                     neutral_point=-20e-3
                 ),
                 gap=4e-3,
-                disk_element_count=DISK_ELEMENT_COUNT,
+                disk_element_count=15,
             ),
 
         ]
@@ -72,9 +70,11 @@ def test_solve_round_constricted_box_equidistant(tmp_path: Path, caplog):
         pass
 
 
-def test_solve_round_constricted_box_uniform(tmp_path: Path, caplog):
+def test_solve_round_constricted_box_uniform(tmp_path: Path, caplog, monkeypatch):
+
     caplog.set_level(logging.INFO, logger="pyroll")
-    pyroll.pillar_model.Config.PILLAR_TYPE = "UNIFORM"
+    monkeypatch.setenv("PILLAR_TYPE", "UNIFORM")
+    monkeypatch.setenv("PILLAR_COUNT", 30)
 
     in_profile = Profile.round(
         diameter=19.5e-3,
@@ -105,7 +105,7 @@ def test_solve_round_constricted_box_uniform(tmp_path: Path, caplog):
                     neutral_point=-20e-3
                 ),
                 gap=4e-3,
-                disk_element_count=DISK_ELEMENT_COUNT,
+                disk_element_count=15,
             ),
 
         ]
