+++ tests/test_solve_round_flat.py
@@ -11,13 +11,11 @@ def pillar_spreads(self: RollPass.DiskElement):
     return self.pillar_draughts ** -0.5
 
 
-DISK_ELEMENT_COUNT = 15
-pyroll.pillar_model.Config.PILLAR_COUNT = 30
-
-
-def test_solve_round_flat_equidistant(tmp_path: Path, caplog):
+def test_solve_round_flat_equidistant(tmp_path: Path, caplog, monkeypatch):
     caplog.set_level(logging.INFO, logger="pyroll")
-    pyroll.pillar_model.Config.PILLAR_TYPE = "EQUIDISTANT"
+
+    monkeypatch.setenv("PILLAR_TYPE", "EQUIDISTANT")
+    monkeypatch.setenv("PILLAR_COUNT", 30)
 
     in_profile = Profile.round(
         diameter=19.5e-3,
@@ -42,7 +40,7 @@ def test_solve_round_flat_equidistant(tmp_path: Path, caplog):
                     neutral_point=-20e-3
                 ),
                 gap=10e-3,
-                disk_element_count=DISK_ELEMENT_COUNT,
+                disk_element_count=15,
             ),
 
         ]
@@ -66,9 +64,11 @@ def test_solve_round_flat_equidistant(tmp_path: Path, caplog):
         pass
 
 
-def test_solve_round_flat_uniform(tmp_path: Path, caplog):
+def test_solve_round_flat_uniform(tmp_path: Path, caplog, monkeypatch):
     caplog.set_level(logging.INFO, logger="pyroll")
-    pyroll.pillar_model.Config.PILLAR_TYPE = "UNIFORM"
+
+    monkeypatch.setenv("PILLAR_TYPE", "UNIFORM")
+    monkeypatch.setenv("PILLAR_COUNT", 30)
 
     in_profile = Profile.round(
         diameter=19.5e-3,
@@ -93,7 +93,7 @@ def test_solve_round_flat_uniform(tmp_path: Path, caplog):
                     neutral_point=-20e-3
                 ),
                 gap=10e-3,
-                disk_element_count=DISK_ELEMENT_COUNT,
+                disk_element_count=15,
             ),
 
         ]
